feat(disk.ts): add K8Disk class for creating a persistent disk
The K8Disk class is a construct that creates a persistent disk for a
Kubernetes cluster. It takes in the size of the disk in GB, the name of
the disk, and the image to use for the disk as properties. If the name
or image properties are not provided, default values are used. The disk
is created using the ComputeDisk resource from the
@cdktf/provider-google library. The type of disk used is pd-ssd.