feat(instance.ts): add variables for instance name, machine type, and boot disk size and create ComputeInstance resource and output its IP address
The instance.ts file now has variables for instance name, machine type,
and boot disk size. A ComputeInstance resource is created with the
specified variables and added to the Terraform stack. The IP address of
the instance is outputted as a Terraform output. This change allows for
more flexibility in creating instances with different configurations and
outputs the IP address for use in other resources.