feat(k8s.ts): add conditional GcsBackend resource creation based on environment variable
The K8Stack constructor now accepts an optional boolean parameter
is_remote, which determines whether to create a GcsBackend resource. If
is_remote is true, then a GcsBackend resource is created, otherwise it
is not. The main.ts file now checks for the presence of the "REMOTE"
environment variable, and if it exists and is truthy, then the K8Stack
constructor is called with is_remote set to true, otherwise it is called
with is_remote set to false. This allows for more flexibility in the
deployment of the K8Stack resource.