feat(main.ts): add yargs for command line argument parsing
The yargs package is added to the project to parse command line
arguments. Two options are added, `remote` and `nodes`. The `remote`
option is a boolean that specifies whether the remote gcs backend will
be used. The `nodes` option is a number that specifies the number of VM
instances for kubernetes nodes.