refactor(k8s.ts): remove numOfNodes variable and use options object instead
The numOfNodes variable has been removed from the K8Stack class and is
now passed as an option object to the constructor. This makes the code
more flexible and easier to read. The main.ts file now passes an options
object to the K8Stack constructor instead of a boolean value.