feat(main.ts): add support for service account credentials file for google provider
The new `--credentials` option allows users to specify the service
account credentials file for the google provider. This is useful when
running the application in a cloud environment where the credentials are
stored in a file. The `K8Stack` constructor now accepts a `credentials`
parameter which is passed to the `GoogleProvider` constructor.