feat(command_options.ts): add command line options for kubernetes cluster deployment
This commit adds command line options for the deployment of a kubernetes
cluster. The options include the GCS bucket name and prefix where the
terraform remote state is stored, whether the remote GCS backend will be
used, the number of VM instances for kubernetes nodes, and the service
account credentials file for the Google provider.