feat(k8stack.ts): add K8Stack class to define a Kubernetes cluster infrastructure
The K8Stack class defines a Kubernetes cluster infrastructure using
Terraform. It creates a VPC network, a master node, and a specified
number of worker nodes. The class also defines variables that can be
used to customize the infrastructure. The infrastructure can be deployed
locally or remotely using Google Cloud Storage as a backend.