feat(k0stack.ts): add K0Stack class
The K0Stack class is a TerraformStack that can be used to create a
Kubernetes cluster. It takes in options such as whether to use a remote
backend, the number of nodes, and the name and prefix of the GCS bucket
to use for remote state. If the remote option is set to true, it creates
a GcsBackend and a DataTerraformRemoteStateGcs to store the remote
state. It also logs the master node IP address.