feat(k8stack.ts): add sshKeyFile property and read_key_file method
The sshKeyFile property is added to the K8StackProperties interface to
allow the user to specify the path to an SSH key file. The read_key_file
method is added to read the contents of the SSH key file and return it
in the format required by the Terraform provider.