feat(create-pull-request.yml): add GitHub Actions workflow to create pull requests from feature branches to develop branch
This commit adds a new GitHub Actions workflow that creates pull
requests from feature branches to the develop branch. The workflow is
triggered on the `create` event and runs on the `ubuntu-20.04` runner.
The `if` condition excludes the `dependabot[bot]` user from triggering
the workflow. The workflow checks out the repository, gets the branch
name, and creates a pull request with the title "Pulling <branch-name>
into develop". The `github_token` secret is used to authenticate the
pull request creation.