feat(.kodiak.toml): add Kodiak configuration file
This commit adds a new file named `.kodiak.toml` which contains
configuration settings for Kodiak. Kodiak is a GitHub bot that automates
various tasks such as merging pull requests and updating dependencies.
The configuration file specifies that pull requests opened by the
dependabot user that upgrade dependencies to minor or patch versions
should be automatically merged. It also specifies that the dependabot
user should be ignored when updating and closing stale dependency
upgrades.