docs(README.md): add instructions for running commands
This commit adds instructions for running commands in the README.md
file. It also provides information on the prerequisites required to run
the commands. The commands are defined in the script field of the
package.json file. The two main commands are `yarn synth` and `yarn
deploy`. To get a list of all available command line options, run `yarn
tsx src/main.ts -h`.