feat: add k0s sample configuration file
This commit adds a sample configuration file for k0s, a lightweight
Kubernetes distribution. The configuration file specifies a cluster with
two hosts, one controller and one worker. The controller has an IP
address of 10.0.0.1 and the worker has an IP address of 10.0.0.2. The
k0s version used is 1.26.0+k0s.0 and dynamic configuration is enabled.