feat(k0s_cluster.ts): add support for enabling cloud provider on worker nodes
This change adds support for enabling cloud provider on worker nodes. A
new function `addCloudProvider` is added to create a YAML node that
contains the necessary flags to enable cloud provider. The
`createHostNodes` function is updated to add the `addCloudProvider` node
to the worker nodes if `enableCloudProvider` is set to true and the node
is a worker.