feat(create_microk8s_bootstrap.ts): add script to create a microk8s bootstrap instance using cdktf
This script creates a microk8s bootstrap instance using the cdktf
library. It takes in various command line arguments such as the Google
Cloud project ID, region, zone, logging bucket name, and more. These
arguments are used to configure the instance and set up the necessary
resources. The script uses the BootStrapInstanceStack class to create
the instance and pass in the arguments.