chore(package.json): add ssh2 and winston packages as dependencies
The ssh2 package is added as a dependency to enable the application to
connect to remote servers via SSH. The winston package is added as a
dependency to enable logging in the application. The @types/ssh2 package
is added as a development dependency to provide type definitions for the
ssh2 package.