feat(install_microk8s.ts): add script to install microk8s on a remote server via ssh
This commit adds a script to install microk8s on a remote server via
ssh. The script takes in command line arguments such as the remote
server host, user name, private ssh key file, and the script to execute
on the remote server. The script logs the output of the installation
process via winston logger.