feat(command_line.ts): add support for certificate template file to be specified via command line argument
The `ct` command line argument is added to specify the certificate
template file that will be copied to the server for IP-based kubectl
access. The default value is set to `templates/csr.conf.template`. The
`csr.conf.template` file is updated to support dynamic IP addresses. The
IP addresses are now specified via the `it.host` variable, which can be
dynamically set.