feat(install_microk8s.ts): add support for uploading a file to the remote server
This commit adds support for uploading a file to the remote server. The
file is uploaded using the SFTP protocol and the `uploadFile` function.
The function takes the source and destination paths as arguments and
logs the progress using the `logger` object. The uploaded file is then
executed on the remote server using the `commandExec` function. This
feature improves the functionality of the `install_microk8s` script by
allowing users to upload and execute files on the remote server.