feat(command_line.ts): add options to remove microk8s and restart script
The command line options now include the ability to remove microk8s and
specify a script to do so. The options are `--remove-script` or `-rms`
for the script path and `--remove-microk8s` or `-rmk` to specify whether
or not to remove microk8s. Additionally, the `--restart-script` or `-rs`
option was already present and allows specifying a script to restart the
cluster.