feat(install_microk8s.ts): add support for remote file upload and execution
This commit adds support for uploading a remote file to the server and
executing it. The new functionality is controlled by the `--rmk` flag,
which specifies the path to the remote file to be uploaded. If the flag
is present, the file is uploaded to the server and executed. The
`uploadFile` function is used to upload the file, and `commandExec` is
used to execute it.