feat(create_cluster.ts): add functionality to create a kops cluster with specified parameters
This commit adds a new file `create_cluster.ts` which contains a
function `runKops` that creates a kops cluster with the specified
parameters. The function takes in command line arguments and constructs
a kops command to create a cluster with the specified parameters. The
`runCommand` function is used to execute the kops command. The function
logs debug and error messages using the `winston` logger.