fix(command_line.ts): change default value of mc option from 2 to 1
The default value of the mc option has been changed from 2 to 1 to
reflect the actual number of kubernetes master nodes that will be
created. The default value of the image option has been changed to
ubuntu-os-cloud/ubuntu-minimal-2204-jammy-v20230428 to reflect the
actual image that will be used for the vm. The default value of the kv
option has been changed to 1.24.12 to reflect the actual version of the
kubernetes cluster.