feat(create_nginx_ingress.ts): add script to create nginx ingress using Terraform and Helm
This script creates an nginx ingress using Terraform and Helm. It uses
the @cdktf/provider-helm library to deploy the ingress-nginx chart to a
Kubernetes cluster. The script takes in several command line arguments
to configure the deployment, including the Kubernetes namespace, chart
name, chart repository, and chart version. It also supports using a
remote GCS backend to store the Terraform state. The script creates a
new Terraform stack and deploys the chart using the HelmProvider and
Release constructs.