feat(install_pgo.ts): add script to install PostgreSQL Operator using Helm chart
This commit adds a new file `install_pgo.ts` which contains a script to
install the PostgreSQL Operator using a Helm chart. The script uses
`cdktf` to create the necessary infrastructure resources. It takes in
several command line arguments to configure the installation, such as
the name of the install, the Kubernetes namespace where the operator
will be installed, the name of the chart, the repository location of the
chart, the Kubernetes config file, the service account credentials file
for Google provider, the GCS bucket name where Terraform remote state is
stored, the GCS bucket folder prefix where Terraform remote state is
stored, whether the remote GCS backend will be used, and the version of
the chart to be deployed.