feat(install_arangodb_operator.ts): add script to install ArangoDB operator via Helm chart
This script installs the ArangoDB operator via a Helm chart. It takes in
several command line arguments to configure the installation, such as
the Kubernetes namespace where the operator will be installed, the Helm
chart repository location, the Kubernetes config file, and the version
of the chart to be deployed. The script also creates a new
`HelmChartStack` construct and passes in the necessary arguments to
deploy the operator.