feat(install_nats.ts): add script to install nats using Helm chart
This commit adds a new script to install nats using a Helm chart. The
script takes in several command line arguments to configure the
installation, including the name of the nats install, the kubernetes
namespace where the nats operator will be installed, the name of the
chart, the repository location of the nats helm chart, the kubernetes
config file, the service account credentials file for the google
provider, the GCS bucket name where terraform remote state is stored,
whether the remote GCS backend will be used, the version of the chart to
be deployed, and the nats version. The script uses the cdktf library to
create a new HelmChartStack and deploys the nats chart with the
specified configuration.