feat(install_postgres_instance.ts): add support for GCS backup credentials
The PostgresSecretStack constructor now takes in a new argument, bc,
which stands for GCS backup credentials. This allows the user to specify
the credentials for backing up the database to GCS. The credentials are
passed to the stack's resource object, which is then used to create a
new GCS key. This feature improves the flexibility of the application as
it allows the user to specify the credentials for backing up the
database to GCS.