feat(install_postgres_instance.ts): add support for backup bucket to store backups
This commit adds a new command line argument `--backup-bucket` (or `-bb`
for short) to the `install_postgres_instance.ts` script. This argument
specifies the name of the GCS bucket where backups will be stored. The
backup bucket is passed to the `PostgresStack` resource as a new
property `backupBucket`. Additionally, the `PostgresSecretStack` is now
instantiated and its `secret` property is passed to the `PostgresStack`
resource.