feat(remote.ts): add RemoteStack class to create a Terraform stack with a GCS backend
The RemoteStack class is a Terraform stack that creates a GCS backend.
It takes in credentials, bucketName, and bucketPrefix as options. The
GcsBackend construct is used to create the backend and the credentials
are read from a file. This class can be used to create a remote backend
for Terraform state.