feat(postgres.ts): add support for multiple repositories
The `repository` field has been added to the `Resource` type and
`PostgresSecretStackProperties` interface. The `repository` field has
also been added to the `PostgresStack` constructor options and
`PostgresSecretStack` constructor options. The `repository` field is
used to specify the name of the repository. The `pgbackrest` object now
uses the `repository` field to set the `repo1-path` and `repo1` fields.
This allows for multiple repositories to be used.