feat(install_postgres_instance.ts): add support for specifying a repository name for backing up the database with pgrest
The `rp` option has been added to the command line arguments to allow
the user to specify a repository name for backing up the database with
pgrest. This allows for more flexibility in the backup process and
allows the user to choose a repository that best suits their needs. The
repository name is passed to the `PostgresSecretStack` and
`PostgresStack` resources as a `resource` property.