feat(create_gcs_bucket.ts): add script to create a GCS bucket for terraform remote state
This script creates a GCS bucket for storing terraform remote state. It
uses the `BucketStack` construct to create the bucket. The script takes
in several command line arguments such as the name of the bucket, the
project id, the region, and the zone. It also allows for the use of a
service account credentials file for the google provider.