feat(helm.ts): add support for listValues in HelmChartProperties to allow setting list values in Helm charts
The HelmChartProperties type now includes a new optional property called
listValues, which is an array of ReleaseSetListStruct. This allows users
to set list values in Helm charts. The HelmChartStack class has been
updated to accept the new listValues property and pass it to the
HelmProvider when creating a new release.