refactor(install_baserow.ts): rename 's' alias to 'sr' for better readability
The 's' alias has been renamed to 'sr' to improve readability. The
publicBackendUrl and publicFrontendUrl options have been added to the
HelmChartStack options to allow for the configuration of the public
URLs. The BaserowIngressStack resource properties have been updated to
use the command line arguments for the namespace, secret,
frontendService, and backendService properties.