feat(create_restic_repo.ts): add script to create a Restic repository stack
The create_restic_repo.ts script is added to the project. This script
uses yargs to parse command line arguments and creates a Restic
repository stack using the RepositoryStack construct from the
../construct/restic module. The script takes various options such as the
name of the installation, the Kubernetes namespace, the Kubernetes
config file, the GCS bucket name for Terraform remote state, whether to
use a remote GCS backend, the GCS bucket for backup, the dictycr secret
containing the Restic credentials, and the image to use for the backup
job. The script creates an instance of the App class, constructs the
deployName using the provided options, and creates a new instance of the
RepositoryStack with the necessary configuration options. Finally, the
app.synth() method is called to synthesize the stack. This script is
added to enable the creation of a Restic repository stack for backup
purposes.