feat(create_restic_pg_backup.ts): add script to create a Restic-based PostgreSQL backup stack
The create_restic_pg_backup.ts script is added to the project. This
script allows users to create a Restic-based PostgreSQL backup stack in
a Kubernetes cluster. It takes in various command line arguments such as
the name of the installation, the Kubernetes namespace, the Kubernetes
config file, the GCS bucket name for storing Terraform remote state, and
other configuration options related to the backup process.