feat(arangodb.ts): add support for creating a database using a Kubernetes Job
The ArangodbSingleStack class has been refactored to include a new
DatabaseStack class. The DatabaseStack class allows for the creation of
a database using a Kubernetes Job. This is achieved by adding a new
import statement for the Job class from the
"@cdktf/provider-kubernetes/lib/job" module. The DatabaseStack class
takes in a set of properties including the namespace, image, adminUser,
adminPassword, user, grant, and databases to configure the database
creation. The #containers method is added to generate the container
configuration for the Job, and the #template_metadata method is added to
generate the metadata for the Job template.