feat(deploy): add script to install ArangoDB database
This commit adds a new file `install_arangodb_database.ts` which contains a script to install an ArangoDB database. The script uses yargs to parse command line arguments and accepts various options such as the name of the installation, the Kubernetes namespace to install, the Kubernetes config file, service account credentials file for GCS backend, GCS bucket name for storing Terraform remote state, whether to use a remote GCS backend, ArangoDB root user, ArangoDB admin user password, ArangoDB user to be created, default database permission for the newly created user, image to use for database and user creations, and the ArangoDB databases to create.

The script creates a new instance of the `App` class from the `cdktf` library and initializes a new `DatabaseStack` with the provided options. The `DatabaseStack` is responsible for creating the necessary resources for the ArangoDB database installation. Finally, the `app.synth()` method is called to synthesize the Terraform code.

The purpose of this change is to automate the installation of an ArangoDB database by providing a script that can be executed with the necessary options.