feat(create_arangodb_databases.ts): add script to create ArangoDB databases and users
This commit adds a new file `create_arangodb_databases.ts` which
contains a script to create ArangoDB databases and users. The script
uses the `yargs` library to parse command line arguments and accepts
various options such as the name of the installation, the Kubernetes
namespace, the Kubernetes config file, the GCS bucket name for Terraform
remote state, and more. The script creates a new instance of the
`DatabaseStack` class from the `../construct/arangodb` module and passes
the parsed arguments to configure the stack. Finally, the script calls
`app.synth()` to generate the CloudFormation template for the stack.
The purpose of this script is to automate the creation of ArangoDB
databases and users during deployment.