feat(install_minio.ts): add script to install Minio using Helm chart
The `install_minio.ts` script is added to the project. This script
allows for the installation of Minio using a Helm chart. It provides
various options such as specifying the name of the Minio install, the
Kubernetes namespace, the chart name, the repository location, the
Kubernetes config file, the service account credentials file, the GCS
bucket name for storing Terraform remote state, whether to use the
remote GCS backend, and the version of the chart to be deployed. The
script creates a new instance of the `HelmChartStack` construct and sets
the necessary configuration options. Finally, the script synthesizes the
CDK app.