docs(README.md): update instructions for service account key file for running terraform
The instructions for setting up the service account key file for running
terraform have been updated. Instead of downloading the key file to the
current folder and renaming it to `credentials.json`, it is now
recommended to download the json formatted service account key and
rename it to something understandable. The key file can be passed to the
`cdktf` scripts using the `[-c|--credentials]` command line option.